requirements.txt
flask~=3.1.2
flask-login~=0.6.3
flask-mysqldb
flask-wtf
werkzeug~=3.1.3
mysqlclient~=2.2.7
protobuf
SQLAlchemy
pandas~=2.3.2
Flask-SQLAlchemy
Pillow
pymysql
PyJWT~=2.10.1
requests~=2.32.5
yarl~=1.20.1
click~=8.2.1
streamlit~=1.49.1
numpy~=2.2.6
matplotlib~=3.10.6
plotly~=6.3.0
cryptography
